requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.0.1
cloudinary==1.31.0
dj-database-url==1.2.0
Django==4.1.6
django-cloudinary-storage==0.3.0
environ==1.0
gunicorn==20.1.0
hiredis==2.2.1
idna==3.4
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pymemcache==4.0.0
python-dotenv==0.21.1
python-magic==0.4.27
redis==4.5.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
whitenoise==6.3.0
